% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cspline}
\alias{cspline}
\title{Cubic Spline Interpolation}
\usage{
cspline(x, y)
}
\arguments{
\item{x,y}{Vectors giving the coordinates of the nodes.}
}
\value{
The cubic spline curve.
}
\description{
Perform cubic spline interpolation with nodes defined by the vectors x and
y, each of length N. The spline is evaluated using the not-a-node end point
conditions (same as Matlab). The y values of the spline curve will be
evaluated at integer points from 0 to x[N-1]. The endpoint x[N-1] is assumed to be an
integer, and the x values are assumed to be in ascending order, with x[0]
equal to 0.
}

