% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecg.R
\docType{data}
\name{ECG}
\alias{ECG}
\title{Electrocardiogram Data}
\format{A time series object.}
\source{
MIT-BIH Normal Sinus Rhythm Database, PhysioBank ATM, \url{http://www.physionet.org/cgi-bin/atm/ATM}
}
\description{
Example ECG data from MIT-BIH Normal Sinus Rhythm Database, ECG1 of record 16265,
 first 2049 observations (0 to 16 seconds with sampling interval of 0.0078125 seconds)
}
\examples{
data(ECG)
plot(ECG)
}
\keyword{datasets}

