% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eemd}
\alias{eemd}
\title{EEMD Decomposition}
\usage{
eemd(input, ensemble_size = 250L, noise_strength = 0.2, S_number = 4L,
  num_siftings = 0L, rng_seed = 0L)
}
\arguments{
\item{input}{Vector of length N. he input signal to decompose.}

\item{ensemble_size}{Number of copies of the input signal to use as the ensemble.}

\item{noise_strength}{Standard deviation of the Gaussian random numbers used as additional
noise. **This value is relative** to the standard deviation of the input signal.}

\item{S_number}{Integer. Use the S-number stopping criterion [2] for the EMD procedure with the given values of `S`.
That is, iterate until the number of extrema and zero crossings in the
signal differ at most by one, and stay the same for S consecutive
iterations. Typical values are in the range 3--8. If \code{S_number} is
zero, this stopping criterion is ignored.}

\item{num_siftings}{Use a maximum number of siftings as a stopping criterion. If
`num_siftings` is zero, this stopping criterion is ignored.}

\item{rng_seed}{A seed for the random number generator. A value of zero denotes
an implementation-defined default value.}
}
\value{
Matrix of size MxN, where M = \code{emd_num_imfs}(N). The rows of the array are the
       IMFs of the input signal, with the last row being the final residual.
}
\description{
Decompose input data to Intrinsic Mode Functions (IMFs) with the
Ensemble Empirical Mode Decomposition algorithm [1].
}
\details{
The size of the ensemble and the relative magnitude of the added noise are
given by parameters \code{ensemble_size} and \code{noise_strength}, respectively.  The
stopping criterion for the decomposition is given by either a S-number or
an absolute number of siftings. In the case that both are positive numbers,
the sifting ends when either of the conditions is fulfilled.
}
\references{
[1] Z. Wu and N. Huang, "Ensemble Empirical Mode Decomposition: A
      Noise-Assisted Data Analysis Method", Advances in Adaptive Data Analysis,
      Vol. 1 (2009) 1–41 \cr
      [2] N. E. Huang, Z. Shen and S. R. Long, "A new view of nonlinear water
      waves: The Hilbert spectrum", Annual Review of Fluid Mechanics, Vol. 31
      (1999) 417–457
}
\seealso{
\code{\link{emd}}
}

