% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extrema}
\alias{extrema}
\title{Find the local extrema of the input data}
\usage{
extrema(x)
}
\arguments{
\item{x}{Vector of length x.}
}
\value{
a list with components ...
}
\description{
Find the local minima and maxima from input data \code{x}. This includes the
artificial extrema added to the ends of the data as specified in the
original EEMD article.
}
\references{
Z. Wu and N. Huang, "Ensemble Empirical Mode Decomposition: A
      Noise-Assisted Data Analysis Method", Advances in Adaptive Data Analysis,
      Vol. 1 (2009) 1–41
}

