% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\docType{data}
\name{float}
\alias{float}
\title{Float Data
 
 The data are a position record from an acoustically tracked subsurface oceanographic float, 
 used as an example data in Rilling et al (2007).}
\format{A time series object.}
\source{
http://wfdac.whoi.edu
}
\description{
Float Data
 
 The data are a position record from an acoustically tracked subsurface oceanographic float, 
 used as an example data in Rilling et al (2007).
}
\examples{
data(float)
plot(float, type = "l"))
}
\keyword{"Bivariate}
\keyword{(2007)}
\keyword{14}
\keyword{936--939}}
\keyword{@references}
\keyword{Decomposition",}
\keyword{Empirical}
\keyword{Flandrin,}
\keyword{Goncalves}
\keyword{IEEE}
\keyword{J.}
\keyword{Letters,}
\keyword{Lilly,}
\keyword{M.}
\keyword{Mode}
\keyword{P.}
\keyword{Processing}
\keyword{Rilling,}
\keyword{Signal}
\keyword{Vol.}
\keyword{\enumerate{}
\keyword{\item{G.}
\keyword{and}
\keyword{datasets}
\keyword{}}
