% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd_num_imfs.R
\name{nIMFs}
\alias{emd_num_imfs}
\alias{nIMFs}
\title{Number of IMFs}
\usage{
emd_num_imfs(N)
}
\arguments{
\item{N}{An integer defining the length of input data.}
}
\value{
The number of IMFs which would be extracted from input data of length N, including
the final residual.
}
\description{
Return the number of IMFs extracted from input data of length N, including
the final residual. This is just [log_2(N)] for N>3.
}

