% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\docType{data}
\name{float}
\alias{float}
\title{Float Data
 
 The data are a position record from an acoustically tracked subsurface oceanographic float, 
 used as an example data in Rilling et al (2007).}
\format{
A time series object.
}
\source{
http://wfdac.whoi.edu
}
\description{
Float Data
 
 The data are a position record from an acoustically tracked subsurface oceanographic float, 
 used as an example data in Rilling et al (2007).
}
\examples{
data("float")
plot(float, type = "l")

}
\references{
\enumerate{
      \item{G. Rilling, P. Flandrin, P. Goncalves and J. M. Lilly,
  "Bivariate Empirical Mode Decomposition",
  IEEE Signal Processing Letters, Vol. 14 (2007) 936--939}
 
}
}
\keyword{datasets}
